move tests to another file
